requirements.txt
aiohttp>=3.9.3
aiosignal>=1.3.1
async-timeout>=4.0.3
attrs>=23.2.0
beautifulsoup4>=4.12.3
Pillow>=10.0.0
numpy>=1.26.4
scikit-learn>=1.4.2
imagehash>=4.3.1
certifi>=2023.1.1
click>=8.1.3
colorama>=0.4.4
frozenlist>=1.4.0
idna>=3.4
joblib>=1.4.0
lxml>=4.9.0
multidict>=6.0.4
nltk>=3.8
peewee>=3.15.0
regex>=2023.1.1
requests>=2.31.0
soupsieve>=2.4.0
tqdm>=4.65.0
typing-extensions>=4.5.0
urllib3>=2.0.0
wrapt>=1.15.0
yarl>=1.9.0
trafilatura>=1.6.0
PyMuPDF>=1.24.4
pytest>=7.0.0
playwright>=1.42.0

# --- Optional ML extras are moved to requirements-ml.txt ---
# For NLI Cross-Encoder + FAISS recalls, also install:
#   pip install -r requirements-ml.txt
